check bam index

+++ maf2igv.py
@@ -57,8 +57,12 @@ def bamlinker(output,bamfile,sample,bamtype):
         if not os.path.lexists(bamL):
             os.symlink(bamfile, bamL)
 
-        stub=bamfile.replace('.bam','')
-        baifile=glob.glob(stub+'*.bai')[0]
+        baifile=bamfile.replace('.bam','.bai')
+        if not os.path.exists(baifile):
+            baifile=bamfile+'.bai'
+        if not os.path.exists(baifile):
+        	print "missing bam index file:", bamfile
+		    raise
         baiL = output + '/' + sample + '.' + bamtype + '.bai'
         if not os.path.lexists(baiL):
             os.symlink(baifile, baiL)
